% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_gam_relationships}
\alias{analyze_gam_relationships}
\title{Analyze GAM relationships}
\usage{
analyze_gam_relationships(time_series_subset)
}
\arguments{
\item{time_series_subset}{A subset of the time series data}
}
\value{
A GAM model object
}
\description{
Fits Generalized Additive Models to capture non-linear relationships.
}
\examples{
\dontrun{
data(pheno_time_series)
subset <- pheno_time_series[pheno_time_series$species == "Quercus_ilex", ]
gam_model <- analyze_gam_relationships(subset)
}
}
