% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos_demo}
\alias{datos_demo}
\title{Datos de demonstration para plasticity analysis}
\format{
Un data.frame con 200 observaciones y 8 variables:
\describe{
\item{insect_population}{Población de insecto (P1, P2, P3, P4)}
\item{fungus_strain}{Cepa de hongo (C1, C2, C3)}
\item{temperature}{Temperatura en °C}
\item{humidity}{Humedad en \%}
\item{insect_size}{Tamaño del insecto en mm}
\item{resistance}{Resistencia del insecto}
\item{fungus_growth}{Crecimiento del hongo}
\item{virulence}{Virulencia del hongo}
}
}
\source{
Simulated data para el paquete phenop
}
\usage{
datos_demo
}
\description{
Dataset simulated que contiene medidas fenotípicas de interacciones insecto-hongo
}
\examples{
data(datos_demo)
head(datos_demo)
}
\keyword{datasets}
