% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_internals.R
\name{.plot_integration}
\alias{.plot_integration}
\title{Funcion interna para grafico de integracion}
\usage{
.plot_integration(multidim_result, traits, plot_options)
}
\arguments{
\item{multidim_result}{Resultado de multidim_plasticity}

\item{traits}{Vector de rasgos}

\item{plot_options}{Opciones de grafico}
}
\value{
Un grafico ggplot
}
\description{
Funcion interna para grafico de integracion
}
\keyword{internal}
