% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{perform_cluster_analysis}
\alias{perform_cluster_analysis}
\title{Perform Cluster Analysis on Phenological Types}
\usage{
perform_cluster_analysis(parameters, n_clusters = 4)
}
\arguments{
\item{parameters}{A data frame containing extracted phenological parameters}

\item{n_clusters}{Number of clusters (default: 4)}
}
\value{
A list with clustering results
}
\description{
Groups populations based on phenological characteristics using k-means.
}
\examples{
\dontrun{
data(pheno_parameters)
clusters <- perform_cluster_analysis(pheno_parameters, n_clusters = 3)
}
}
