% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{perform_pca_analysis}
\alias{perform_pca_analysis}
\title{Perform Principal Component Analysis}
\usage{
perform_pca_analysis(parameters)
}
\arguments{
\item{parameters}{A data frame containing extracted phenological parameters}
}
\value{
A PCA result object
}
\description{
Performs PCA on phenological parameters to identify main axes of variation.
}
\examples{
\dontrun{
data(pheno_parameters)
pca_results <- perform_pca_analysis(pheno_parameters)
}
}
