% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pheno_spatial}
\alias{pheno_spatial}
\title{Spatial Phenological Data}
\format{
A data frame with X rows and 24 variables:
\describe{
\item{site_id}{Unique identifier for the spatial site}
\item{longitude}{Longitude coordinate in decimal degrees}
\item{latitude}{Latitude coordinate in decimal degrees}
\item{elevation}{Elevation above sea level (meters)}
\item{mean_annual_temp}{Mean annual temperature (°C)}
\item{annual_precipitation}{Total annual precipitation (mm)}
\item{soil_type}{Classification of soil type}
\item{soil_pH}{Soil pH measurement}
\item{soil_organic_carbon}{Soil organic carbon content (g/kg)}
\item{sos}{Start of season (day of year)}
\item{eos}{End of season (day of year)}
\item{los}{Length of season (days)}
\item{ndvi_max}{Maximum Normalized Difference Vegetation Index value}
\item{land_cover}{Land cover classification type}
\item{forest_type}{Forest type classification}
\item{protected_area}{Indicates if the site is in a protected area}
\item{years_since_disturbance}{Years since last disturbance event}
\item{disturbance_type}{Type of disturbance (fire, logging, etc.)}
\item{climate_zone}{Climate zone classification}
\item{phenology_class}{Phenology classification category}
\item{productivity}{Productivity metric value}
\item{productivity_class}{Productivity classification category}
\item{species_richness}{Number of species recorded at the site}
\item{aboveground_carbon}{Aboveground carbon stock (Mg/ha)}
}
}
\usage{
pheno_spatial
}
\description{
Gridded spatial data for phenological mapping and spatial analysis.
}
\keyword{datasets}
