% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasticidad_ejemplo.R
\name{plasticidad_ejemplo}
\alias{plasticidad_ejemplo}
\title{Example dataset para plasticidad fenotipica}
\format{
Un data.frame con 200 observaciones y 6 variables.
}
\source{
Simulado
}
\description{
Simulated data to demonstrate las funciones of the package phenop.
}
