% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_internals.R
\name{plot_internals}
\alias{plot_internals}
\title{Internal plotting function for multidimensional plasticity}
\usage{
plot_internals(df, type = "integration", traits = NULL, plot_options = list())
}
\arguments{
\item{df}{Data frame or result object}

\item{type}{Type of plot: "integration" or "radar"}

\item{traits}{Optional vector of traits}

\item{plot_options}{List of plot options}
}
\value{
ggplot object
}
\description{
Internal plotting function for multidimensional plasticity
}
