% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_pathogen_interaction_extended.R
\name{print.host_pathogen_interaction_extended}
\alias{print.host_pathogen_interaction_extended}
\title{Print method for host_pathogen_interaction_extended}
\usage{
\method{print}{host_pathogen_interaction_extended}(x, ...)
}
\arguments{
\item{x}{An object of class host_pathogen_interaction_extended}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for host_pathogen_interaction_extended
}
