% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasticity_meta_analysis_extended.R
\name{print.plasticity_meta_analysis_extended}
\alias{print.plasticity_meta_analysis_extended}
\title{Print method for plasticity_meta_analysis_extended}
\usage{
\method{print}{plasticity_meta_analysis_extended}(x, ...)
}
\arguments{
\item{x}{An object of class plasticity_meta_analysis_extended}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for plasticity_meta_analysis_extended
}
