% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasticity_tradeoffs_extended.R
\name{print.plasticity_tradeoffs_extended}
\alias{print.plasticity_tradeoffs_extended}
\title{Print method for plasticity_tradeoffs_extended}
\usage{
\method{print}{plasticity_tradeoffs_extended}(x, ...)
}
\arguments{
\item{x}{An object of class plasticity_tradeoffs_extended}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for plasticity_tradeoffs_extended
}
