% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidim_plasticity_extended.R
\name{print.plot_multidim_result}
\alias{print.plot_multidim_result}
\title{Print method for plot objects}
\usage{
\method{print}{plot_multidim_result}(x, ...)
}
\arguments{
\item{x}{A plot object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for plot objects
}
