% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_multidim_plasticity.R
\name{print.safe_multidim_result}
\alias{print.safe_multidim_result}
\title{Print method for safe_multidim_result}
\usage{
\method{print}{safe_multidim_result}(x, ...)
}
\arguments{
\item{x}{An object of class safe_multidim_result}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for safe_multidim_result
}
