% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenop-package.R
\docType{package}
\name{phenop-package}
\alias{phenop}
\alias{phenop-package}
\title{phenop: Multidimensional Analysis of Phenotypic Plasticity in Eco-Evolutionary Contexts}
\description{
Provides novel methods for analyzing phenotypic plasticity across multiple traits and environmental variables, with special focus on insect-fungus host-pathogen systems. Implements original multidimensional plasticity indices and interaction analyses not available in other packages.
}
\author{
\strong{Maintainer}: Leonel Stazione \email{leonelstazione@hotmail.com}

}
\keyword{internal}
